refactor: extract download logic into downloadUtils utility
